Add login screen for root site Public access to "/workshop" and "/public"
TODO: create new public CityPyO User
TODO: Rename "workshop" variable into "publicAccess"
