+++ src/main.ts
@@ -1,5 +1,6 @@
 import Vue from 'vue'
-import App from './App.vue'
+import Home from './Home.vue'
+//import Login from './components/Login/Login.vue'
 import router from './router'
 import store from './store'
 import vuetify from './plugins/vuetify'
@@ -12,5 +13,5 @@ new Vue({
     router,
     store,
     vuetify,
-    render: h => h(App)
+    render: h => h(Home)
 }).$mount('#app')
